feat: add code for service in order-service and proto code

+++ services/order-service/pkg/dto/order_dto.go
@@ -6,7 +6,7 @@ import (
 
 type Order struct {
 	ID         uint64
-	BuyerID    int
+	BuyerID    uint64
 	Status     string
 	TotalPrice int
 	OrderItems []*OrderItem
